requirements.txt
discord.py
async-google-trans-new
Flask
replit
urlib3
